% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{CityWeather}
\alias{CityWeather}
\title{CityWeather class constructor}
\usage{
CityWeather(city, temperature, humidity, pressure)
}
\arguments{
\item{city}{name of the city}

\item{temperature}{numeric temperature (e.g., in °C)}

\item{humidity}{numeric humidity in percent (0–100)}

\item{pressure}{numeric pressure in hPa}
}
\value{
an object of class City_Weather
}
\description{
CityWeather class constructor
}
