% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{expected_lambda}
\alias{expected_lambda}
\title{Expected number of guests from weather}
\usage{
expected_lambda(city_weather)
}
\arguments{
\item{city_weather}{City_Weather object}
}
\value{
numeric expected number of guests (lambda)
}
\description{
Uses the link:
  lambda = exp(0.5 + 0.5 T - 3 H + 0.001 P)
where H is rescaled from percent to [0, 1].
}
