% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\name{GlassProfile}
\alias{GlassProfile}
\title{GlassProfile constructor}
\usage{
GlassProfile(a, b, x_1, x_2, x_3, x_4, r_foot, r_stem, r_bowl, r_rim)
}
\arguments{
\item{a, b}{numeric limits for champagne level}

\item{x_1, x_2, x_3, x_4}{breakpoints along the glass}

\item{r_foot, r_stem, r_bowl, r_rim}{radii at key points}
}
\value{
an object of class Glass_Profile
}
\description{
GlassProfile constructor
}
