% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{r}
\alias{r}
\title{Radius profile r(t)}
\usage{
r(t, x_1, x_2, x_3, x_4, r_foot, r_stem, r_bowl, r_rim)
}
\arguments{
\item{t}{position along the glass axis}

\item{x_1, x_2, x_3, x_4}{breakpoints}

\item{r_foot, r_stem, r_bowl, r_rim}{radii}
}
\value{
numeric radius at t
}
\description{
Radius profile r(t)
}
