% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{r_vec_for_integrate}
\alias{r_vec_for_integrate}
\title{Vectorized radius for integration}
\usage{
r_vec_for_integrate(t, ...)
}
\arguments{
\item{t}{numeric vector}

\item{...}{parameters passed to r()}
}
\value{
numeric vector of radius
}
\description{
Vectorized radius for integration
}
