% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\name{radius_cone}
\alias{radius_cone}
\title{radius_cone method}
\usage{
radius_cone(x, t)
}
\arguments{
\item{x}{Glass_Profile object}

\item{t}{numeric position}
}
\value{
radius at t
}
\description{
radius_cone method
}
