% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_parties}
\alias{simulate_parties}
\title{Simulate many champagne parties}
\usage{
simulate_parties(city_weather, glass, N = 1000L, bottle_L = 0.75, seed = 123L)
}
\arguments{
\item{city_weather}{City_Weather object}

\item{glass}{Glass_Profile object}

\item{N}{number of simulated parties}

\item{bottle_L}{bottle size in liters}

\item{seed}{integer random seed}
}
\value{
data.frame with columns guests, glasses, total_L, bottles
}
\description{
Simulate \code{N} parties given weather conditions and a glass shape.
}
\examples{
\dontrun{
glass <- GlassProfile(
  a = 12.6, b = 22.5,
  x_1 = 1.1, x_2 = 12.6, x_3 = 18.8, x_4 = 22.5,
  r_foot = 5.8, r_stem = 0.25, r_bowl = 7.5, r_rim = 7
)

bern <- CityWeather("Bern", temperature = 20, humidity = 60, pressure = 1013)
simulate_parties(bern, glass, N = 10)
}
}
