% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\name{volume_between}
\alias{volume_between}
\title{volume_between method}
\usage{
volume_between(x, lower = x$a, upper = x$b)
}
\arguments{
\item{x}{Glass_Profile object}

\item{lower}{lower integration limit (default: x$a)}

\item{upper}{upper integration limit (default: x$b)}
}
\value{
numeric volume between lower and upper (cm^3)
}
\description{
volume_between method
}
